% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_uk2us}
\alias{convert_uk2us}
\title{Convert words from UK to US English}
\usage{
convert_uk2us(x, crosswalk = uk2us::ukus_crosswalk)
}
\arguments{
\item{x}{Vector of words to convert.}

\item{crosswalk}{Data frame with columns \code{uk} and \code{us}.}
}
\value{
A vector of converted words.
}
\description{
\code{convert_uk2us} converts a vector of words from UK to US English.
}
\examples{
convert_uk2us(c("centre", "colour", "colourize"))

}
\seealso{
\code{\link{convert_us2uk}}
}
