% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_us2uk}
\alias{convert_us2uk}
\title{Convert words from US to UK English}
\usage{
convert_us2uk(x, crosswalk = uk2us::ukus_crosswalk)
}
\arguments{
\item{x}{Vector of words to convert.}

\item{crosswalk}{Data frame with columns \code{uk} and \code{us}.}
}
\value{
A vector of converted words.
}
\description{
\code{convert_us2uk} converts a vector of words from US to UK English.
}
\examples{
convert_us2uk(c("center", "color", "colorize"))

}
\seealso{
\code{\link{convert_uk2us}}
}
