% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukus_crosswalk.R
\docType{data}
\name{ukus_crosswalk}
\alias{ukus_crosswalk}
\title{Crosswalk between UK and US English spellings}
\format{
Data frame with columns
\describe{
\item{uk}{UK English spelling}
\item{us}{US English spelling}
}
}
\source{
\url{http://www.tysto.com/uk-us-spelling-list.html}
}
\usage{
data(ukus_crosswalk)
}
\description{
Data frame containing correspondence between UK and US English spellings.
}
\keyword{datasets}
